requirements.txt
asgiref==3.2.10
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==3.1
django-autoslug==1.9.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
jwt==1.3.1
marshmallow==3.17.0
mysqlclient @ file:///C:/Users/lucas/Documents/Progra/django/deshuezadero/mysqlclient-1.4.6-cp37-cp37m-win_amd64.whl#sha256=683269b350813b3d593e290e0d565b0de6bcd7f8b3c5cec3025260b3dfcd3b57
packaging==23.1
Pillow==9.5.0
pycparser==2.21
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
transbank-sdk==4.0.0
typing_extensions==4.5.0
urllib3==2.0.2
